move project copy stuff to its own module so it doesn't clutter project.rb
Conflicts:
	spec/models/project_spec.rb
